requirements.txt
pypiwin32==223
pycryptodome==3.20.0